+++ packages/google_maps_flutter/google_maps_flutter_android/pigeons/messages.dart
@@ -11,6 +11,15 @@ import 'package:pigeon/pigeon.dart';
   copyrightHeader: 'pigeons/copyright.txt',
 ))
 
+/// Pigeon equivalent of MapType
+enum PlatformMapType {
+  none,
+  normal,
+  satellite,
+  terrain,
+  hybrid,
+}
+
 // Pigeon equivalent of the Java MapsInitializer.Renderer.
 enum PlatformRendererType { legacy, latest }
 
@@ -34,7 +43,7 @@ class PlatformCameraUpdate {
   PlatformCameraUpdate(this.json);
 
   /// The update data, as JSON. This should only be set from
-  /// CameraUpdate.toJson, and the native code must intepret it according to the
+  /// CameraUpdate.toJson, and the native code must interpret it according to the
   /// internal implementation details of the CameraUpdate class.
   // TODO(stuartmorgan): Update the google_maps_platform_interface CameraUpdate
   //  class to provide a structured representation of an update. Currently it
@@ -49,7 +58,19 @@ class PlatformCircle {
   PlatformCircle(this.json);
 
   /// The circle data, as JSON. This should only be set from
-  /// Circle.toJson, and the native code must intepret it according to the
+  /// Circle.toJson, and the native code must interpret it according to the
+  /// internal implementation details of that method.
+  // TODO(stuartmorgan): Replace this with structured data. This exists only to
+  //  allow incremental migration to Pigeon.
+  final Map<String?, Object?> json;
+}
+
+/// Pigeon equivalent of the Heatmap class.
+class PlatformHeatmap {
+  PlatformHeatmap(this.json);
+
+  /// The heatmap data, as JSON. This should only be set from
+  /// Heatmap.toJson, and the native code must interpret it according to the
   /// internal implementation details of that method.
   // TODO(stuartmorgan): Replace this with structured data. This exists only to
   //  allow incremental migration to Pigeon.
@@ -68,7 +89,7 @@ class PlatformMarker {
   PlatformMarker(this.json);
 
   /// The marker data, as JSON. This should only be set from
-  /// Marker.toJson, and the native code must intepret it according to the
+  /// Marker.toJson, and the native code must interpret it according to the
   /// internal implementation details of that method.
   // TODO(stuartmorgan): Replace this with structured data. This exists only to
   //  allow incremental migration to Pigeon.
@@ -80,7 +101,7 @@ class PlatformPolygon {
   PlatformPolygon(this.json);
 
   /// The polygon data, as JSON. This should only be set from
-  /// Polygon.toJson, and the native code must intepret it according to the
+  /// Polygon.toJson, and the native code must interpret it according to the
   /// internal implementation details of that method.
   // TODO(stuartmorgan): Replace this with structured data. This exists only to
   //  allow incremental migration to Pigeon.
@@ -92,7 +113,7 @@ class PlatformPolyline {
   PlatformPolyline(this.json);
 
   /// The polyline data, as JSON. This should only be set from
-  /// Polyline.toJson, and the native code must intepret it according to the
+  /// Polyline.toJson, and the native code must interpret it according to the
   /// internal implementation details of that method.
   // TODO(stuartmorgan): Replace this with structured data. This exists only to
   //  allow incremental migration to Pigeon.
@@ -113,11 +134,26 @@ class PlatformTileOverlay {
   PlatformTileOverlay(this.json);
 
   /// The tile overlay data, as JSON. This should only be set from
-  /// TileOverlay.toJson, and the native code must intepret it according to the
+  /// TileOverlay.toJson, and the native code must interpret it according to the
   /// internal implementation details of that method.
   // TODO(stuartmorgan): Replace this with structured data. This exists only to
   //  allow incremental migration to Pigeon.
   final Map<String?, Object?> json;
+}
+
+/// Pigeon equivalent of Flutter's EdgeInsets.
+class PlatformEdgeInsets {
+  PlatformEdgeInsets({
+    required this.top,
+    required this.bottom,
+    required this.left,
+    required this.right,
+  });
+
+  final double top;
+  final double bottom;
+  final double left;
+  final double right;
 }
 
 /// Pigeon equivalent of LatLng.
@@ -154,16 +190,89 @@ class PlatformCluster {
   final List<String?> markerIds;
 }
 
+/// Pigeon equivalent of CameraTargetBounds.
+///
+/// As with the Dart version, it exists to distinguish between not setting a
+/// a target, and having an explicitly unbounded target (null [bounds]).
+class PlatformCameraTargetBounds {
+  PlatformCameraTargetBounds({required this.bounds});
+
+  final PlatformLatLngBounds? bounds;
+}
+
+/// Information passed to the platform view creation.
+class PlatformMapViewCreationParams {
+  PlatformMapViewCreationParams({
+    required this.initialCameraPosition,
+    required this.mapConfiguration,
+    required this.initialCircles,
+    required this.initialMarkers,
+    required this.initialPolygons,
+    required this.initialPolylines,
+    required this.initialHeatmaps,
+    required this.initialTileOverlays,
+    required this.initialClusterManagers,
+  });
+
+  final PlatformCameraPosition initialCameraPosition;
+  final PlatformMapConfiguration mapConfiguration;
+  // TODO(stuartmorgan): Make the generic types non-nullable once supported.
+  // https://github.com/flutter/flutter/issues/97848
+  // The consuming code treats the entries as non-nullable.
+  final List<PlatformCircle?> initialCircles;
+  final List<PlatformMarker?> initialMarkers;
+  final List<PlatformPolygon?> initialPolygons;
+  final List<PlatformPolyline?> initialPolylines;
+  final List<PlatformHeatmap?> initialHeatmaps;
+  final List<PlatformTileOverlay?> initialTileOverlays;
+  final List<PlatformClusterManager?> initialClusterManagers;
+}
+
 /// Pigeon equivalent of MapConfiguration.
 class PlatformMapConfiguration {
-  PlatformMapConfiguration({required this.json});
+  PlatformMapConfiguration({
+    required this.compassEnabled,
+    required this.cameraTargetBounds,
+    required this.mapType,
+    required this.minMaxZoomPreference,
+    required this.mapToolbarEnabled,
+    required this.rotateGesturesEnabled,
+    required this.scrollGesturesEnabled,
+    required this.tiltGesturesEnabled,
+    required this.trackCameraPosition,
+    required this.zoomControlsEnabled,
+    required this.zoomGesturesEnabled,
+    required this.myLocationEnabled,
+    required this.myLocationButtonEnabled,
+    required this.padding,
+    required this.indoorViewEnabled,
+    required this.trafficEnabled,
+    required this.buildingsEnabled,
+    required this.liteModeEnabled,
+    required this.cloudMapId,
+    required this.style,
+  });
 
-  /// The configuration options, as JSON. This should only be set from
-  /// _jsonForMapConfiguration, and the native code must intepret it according
-  /// to the internal implementation details of that method.
-  // TODO(stuartmorgan): Replace this with structured data. This exists only to
-  //  allow incremental migration to Pigeon.
-  final Map<String?, Object?> json;
+  final bool? compassEnabled;
+  final PlatformCameraTargetBounds? cameraTargetBounds;
+  final PlatformMapType? mapType;
+  final PlatformZoomRange? minMaxZoomPreference;
+  final bool? mapToolbarEnabled;
+  final bool? rotateGesturesEnabled;
+  final bool? scrollGesturesEnabled;
+  final bool? tiltGesturesEnabled;
+  final bool? trackCameraPosition;
+  final bool? zoomControlsEnabled;
+  final bool? zoomGesturesEnabled;
+  final bool? myLocationEnabled;
+  final bool? myLocationButtonEnabled;
+  final PlatformEdgeInsets? padding;
+  final bool? indoorViewEnabled;
+  final bool? trafficEnabled;
+  final bool? buildingsEnabled;
+  final bool? liteModeEnabled;
+  final String? cloudMapId;
+  final String? style;
 }
 
 /// Pigeon representation of an x,y coordinate.
@@ -193,8 +302,8 @@ class PlatformTileLayer {
 class PlatformZoomRange {
   PlatformZoomRange({required this.min, required this.max});
 
-  final double min;
-  final double max;
+  final double? min;
+  final double? max;
 }
 
 /// Interface for non-test interactions with the native SDK.
@@ -218,6 +327,13 @@ abstract class MapsApi {
   // The consuming code treats the entries as non-nullable.
   void updateCircles(List<PlatformCircle?> toAdd,
       List<PlatformCircle?> toChange, List<String?> idsToRemove);
+
+  /// Updates the set of heatmaps on the map.
+  // TODO(stuartmorgan): Make the generic type non-nullable once supported.
+  // https://github.com/flutter/flutter/issues/97848
+  // The consuming code treats the entries as non-nullable.
+  void updateHeatmaps(List<PlatformHeatmap?> toAdd,
+      List<PlatformHeatmap?> toChange, List<String?> idsToRemove);
 
   /// Updates the set of custer managers for clusters on the map.
   // TODO(stuartmorgan): Make the generic type non-nullable once supported.
@@ -367,6 +483,15 @@ abstract class MapsInitializerApi {
   @async
   PlatformRendererType initializeWithPreferredRenderer(
       PlatformRendererType? type);
+}
+
+/// Dummy interface to force generation of the platform view creation params,
+/// which are not used in any Pigeon calls, only the platform view creation
+/// call made internally by Flutter.
+@HostApi()
+abstract class MapsPlatformViewApi {
+  // This is never actually called.
+  void createView(PlatformMapViewCreationParams? type);
 }
 
 /// Inspector API only intended for use in integration tests.